Remove the need for FUSE by extracting the AppImage
For the AppImage contents to work correctly, the files contained
in the AppImage must be accessible at their absolute paths
(that's what the AppImage runtime does, overlaying the FUSE-mounted
filesystem over the root FS).

Since we're building a Unity-specific docker image, there is no
harm in "permanently" copying those files at the root of the image's
filesystem. This then makes the AppImage's entrypoint `AppRun` work.